2_20 Yiyang Zeng
Hey Guys, I have made a few changes to the website:                                     1. I put all the search criteria on the search/recommendation page.                      2. I combine two search result tables into one.                                                   3. I changed the search criteria to more reasonable ones.                               4. I added the Strength and the Weakness columns to show the hospital's strength and weakness. (where they are above/below the national average)                                                                                    5. I enable the user to enter a zipcode, and the website will show which city and state of the user is in. Then, the website also calculates the distance between hospitals and the user.                                                                        6. I also get rid of the previous ordering algorithm. I use the number of strength, the number of weakness and the cost to somehow calculate the ranks.